+++ YouFeel/Module/Home/MyIntegral/Controllers/BXTRankingViewController.m
@@ -67,7 +67,7 @@
     }
     else
     {
-        cell.rankingView.text = [NSString stringWithFormat:@"%ld", indexPath.row + 1];
+        cell.rankingView.text = [NSString stringWithFormat:@"%ld", (long)indexPath.row + 1];
     }
     
     return cell;
